add: test deployment
